+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * i : Translations cleanup and move to native translations
 * i : Simply notify if permission are missing (permission to add using browser interface)
 * i : UserScript - Replace context menu with panel button to refresh
+* i : Replace update check with socket update information
 * \- : Remove `iqdb` search
 * \- : Remove copy text link